Fixes messages between emitter and receiver (#39)
On Talentsoft the messages sended by the app are not received by the LMS.
It seems that we are not sending messages to the correct window, I don't know exactly why because it worked by the past.

To fix it, I used `window.top` to be sure that we are always sending to the same window.