SUPPORT-1259 Make sure scorm connect setLessonStatus always commit after lesson status change (#59)
[Contributing Guidelines](https://github.com/Didask/odc#contributing)

# ❓Problem

Sometimes scorm connect users close badly the LMS tab and changes are not committed (progression not saved to the LMS).

# 🎯 Solution

In scorm connect, force commit each time lesson status is set.
We also want to be sure the score is set before the status to make sure the last score is committed.

# 🧭 Tests

- [x] Test with scorm cloud

# 🧑‍🎓 Checklist

- [x] I have reviewed my own PR